+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to MemWord Project for Computer Programming II Class.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0] - 2018-12-11
+
+### Added
+- Search/Delete feature in dictionary selection GUI
+- Simple statistics panel for checking study progress
+- Full-implementation of Quiz GUI
+- Inter-movement of all GUIs
+- Various DB methods for new features
+
+### Changes
+- DB code revision
+- GUI layout to GridBagLayout
+- Divided Quiz GUI frame and their panels
+
 ## [0.5] - 2018-12-05
 ### Added
 - Finished basic features of Learn GUI (prompt added)
@@ -47,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parsing csv files into DB
 - Querying data from DB
 
-[0.5]: https://github.com/MariAli-Lover/MemWord/compare/595394a...master
+[0.6]: https://github.com/MariAli-Lover/MemWord/compare/87bba6a...master
+[0.5]: https://github.com/MariAli-Lover/MemWord/compare/595394a...87bba6a
 [0.4]: https://github.com/MariAli-Lover/MemWord/compare/a04cf28...595394a 
 [0.3]: https://github.com/MariAli-Lover/MemWord/compare/8ee2bae...a04cf28
 [0.2]: https://github.com/MariAli-Lover/MemWord/compare/a8b9ac2...8ee2bae